chore: consolidate http/https fetching
